Refactored module to index.js to avoid duplicating names in require calls
