+++ index.js
@@ -1,0 +1,7 @@
+const path = require('path');
+
+const bn128 = require(path.join(__dirname, 'bn128'));
+
+module.exports = {
+    bn128
+};
