+++ keccak/index.js
