+++ setup/index.js
@@ -15,7 +15,7 @@ const path = require('path');
 
 const setup = {};
 const { SIGNATURES_PER_FILE } = require('../params');
-const bn128 = require('../bn128/bn128');
+const bn128 = require('../bn128');
 
 const partialPath = path.posix.resolve(__dirname, '../setupDatabase');
 const compressionMask = new BN('8000000000000000000000000000000000000000000000000000000000000000', 16);
