+++ test/proof/extractor.spec.js
@@ -4,8 +4,8 @@ const crypto = require('crypto');
 const { padLeft } = require('web3-utils');
 const sinon = require('sinon');
 
-const bn128 = require('../../bn128/bn128');
-const proof = require('../../proof/proof');
+const bn128 = require('../../bn128');
+const proof = require('../../proof');
 const extractor = require('../../proof/extractor');
 const proofHelpers = require('../../proof/helpers');
 const { K_MAX } = require('../../params');
