+++ test/proof/verifier.spec.js
@@ -5,8 +5,8 @@ const crypto = require('crypto');
 const { padLeft, sha3 } = require('web3-utils');
 const sinon = require('sinon');
 
-const bn128 = require('../../bn128/bn128');
-const proof = require('../../proof/proof');
+const bn128 = require('../../bn128');
+const proof = require('../../proof');
 const verifier = require('../../proof/verifier');
 const proofHelpers = require('../../proof/helpers');
 const { K_MAX } = require('../../params');
